Moved the solarized colors into their own file.
This also sets up the ability to use light or dark colors and whether
or not to tailor the colors to 256 term friendly ones.

Instead of battling with the bad contrast of bright black to the
background color, I just made both blacks the same color, deciding to
give up on brightblack in favor of something readable and easy.
Going to try it out for a while to see if that gels for me.