Add flake for WezTerm
WezTerm hasn't done a new tagged release in a long time, so lots of bug
fixes that they have made aren't in nixpkgs. So adding a flake to get
the latest WezTerm - for the hosts that need it.

The hosts which have NVidia GPUs seem to need the update in order to not
have WezTerm crash all the time. Also with the new update, Wayland is
supported again, so that can be turned on.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>