Update flake
Needed to use the stable version of the Rust VS Code extension. The VS
Code file was also updated based on the new setting location for
extensions.

Lastly, removed the Firefox speech synthesis options out of home-manager
and into the system nixos configs. This is because home-manager doesn't
like it when `useGlobal` is set, and then the global values are
modified.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>