Add the latest tytools
Version 0.9.9 was submitted, but there were MacOS issues in the CI, so
it was rolled back. I took that PR file and just added it to my packages
so that I can use it.

Version 0.9.9 is needed for the M8, otherwise it craps out when trying
to update the firmware.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>