Reduce redundancy in the gtk theme setup and allow for other themes
Added an explicit theme setting to GTK so that I can add additional
themes that are neutral and not particularly tied to a color theme. As I
did so, found opportunities to add attributes to the default module so
that each theme didn't have to define it themselves. This wasn't done
before because I didn't realize I could just reference the `config` as a
setting and thought it needed to be a scoped variable here or there.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>