Rework the MPV module
Now swappable parts of the code are based on profiles, which in turn
contain all shaders and now parameters like scaling in them.

The defs file contains the definition of each profile - which can then
be turned into MPV profiles, along with the bindings for that
definition.

As a result of this re-work, I was able to remove a bunch of packages
that I didn't need anymore - either because they were unused or because
they could be found in the default shaders package.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>