Add umport lib for recursivly finding package files
What this allows me to do is have a single directory with multiple nix
files/packages in them to load. This means, all the fonts can go in one
fonts directory.

This is pre-work to start pulling out some of the scripts that exist in
various modules into the packages directory. This allows me to build
them individually without having to rebuild the whole nixos config.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>