Read the packages to automatically add to overlay
Did this so that anytime a new package is added, I don't have to
remember to update the overlays code as well. But it has some
compromises.

First, The naming convention of the packages matters - as it's used for
determining the package name. I was using `pname` or `name` from the
result of `callPackage`; but that doesn't work for overlays. The reason
is that I try to evaluate the derivation at all for pulling names, I end
up in an infinite recursion state. Using file names as a convention just
makes it easy to pull that info without evaluating anything.

In the future, I might be able to have a wrapper structure - but for
now, this works.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>