Comment out controlPersist
With the latest OpenSSH update, that setting seems to make the
connection to SSH servers feel like they are hanging. Typing and seeing
the output from the server is very delayed.

Disabling controlPersist seems to address the issue, and the 5 min value
isn't too big of a loss for me right now

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>